NSDecimalNumber replaced with Decimal

+++ Revolut/Domain/RatesList/RatesList.swift
@@ -11,5 +11,5 @@ import Foundation
 struct RatesList: Codable {
     let base: String
     let date: String
-    let rates: [String : Double]
+    let rates: [String : Decimal]
 }
